fetch/metadata/generated/worker-dedicated-constructor-sec-fetch-dest.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <script type="module">
  'use strict';
  function induceRequest(path, options) {
    const url = path + '&body=postMessage("")&mime=application/javascript';
    return new Promise((resolve, reject) => {
      const worker = new Worker(url, options);
      worker.onmessage = resolve;
      worker.onerror = reject;
    });
  }

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';

    return induceRequest('https://{{host}}:{{ports[https][0]}}' + path)
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'worker');
        });
  }, 'sec-fetch-dest - classic script');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';

    return induceRequest('https://{{host}}:{{ports[https][0]}}' + path, {"type": "module"})
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'worker');
        });
  }, 'sec-fetch-dest - module script');
  </script>
</html>
